requirements.txt
Django==2.0.6
psycopg2-binary==2.7.4
django-admin-view-permission==1.6
gunicorn
django-admin-list-filter-dropdown
django-modeladmin-reorder
django-compressor==2.2
csscompressor==0.9.5
django-templated-email==2.2.0
django-nested-admin==3.1.2
docker-compose
coverage
